+++ Assets/Scripts/BattleScene/Enemy's/Enemy.cs
@@ -51,7 +51,7 @@ public class Enemy : BasicUnitScript
         StartCoroutine(Attacking(true, nowAttackCount_I, 1f)); //첫번째 공격 실행
     }
 
-    public override void Hit(int damage, bool isDefending)
+    public override void Hit(float damage, bool isDefending)
     {
         base.Hit(damage, isDefending);
     }
